// src/types/enums.ts
export enum COLOR_CODE {
  PRIMARY = '#B48442',
  COMPLEMENTARY = '#D6A561',
}

export enum BACKGROUND_COLOR {
  PRIMARY = '#FCF6F0',
  WHITE = '#FFF',
  BLACK = '#000',
  BLACK1 = '#1D1E18',
  GRAY_2 = '#333',
  LIGHT_GOLD = '#F3BC6F',
}

export enum TEXT_COLOR {
  WHITE = '#FFF',
  BLACK_1 = '#1D1E18',
  BLACK = '#000',
  GRAY = '#676767',
  GRAY1 = '#808080',
  GRAY2 = '#9C9C9C',
  GRAY3 = '#5C5C5C',
  DARK_BLUE = '#202F60',
}

export enum COLOR_PALETTE {
  SEASHELL = '#FCF6F0',
  MARIGOLD = '#B48442',
  MIDNIGHT = '#16234E',
  NIGHT_RIDER = '#333333',
  APACHE = '#D8A763',
  CHEROKEE = '#FFCE8B',
  BLANCHED_ALMOND = '#FFE8CC',
  HARVEST_GOLD = '#DEB071',
  DRIFTWOOD = '#B58443',
  GREENISH_TEAL = '#3DBEAF',
  JUNGLE_GREEN = '#36C68B',
  TRANSPARENT = 'transparent',
  GRAY = '#333',
  GRAY1 = '#808080',
  GRAY2 = '#9C9C9C',
}

export enum RECIPE_TYPE {
  BY_YOU = 'by_you',
  FOR_YOU = 'for_you',
}

export enum CAROUSEL_NAVIGATION {
  NEXT = 'next',
  PREV = 'prev',
}